requirements.txt
# Python Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.25.0
playwright>=1.40.0
pyyaml>=6.0
loguru>=0.7.0
